requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.4
daphne==4.1.0
Django==5.0.3
django-core==1.4.1
django-cors-headers==4.3.1
django-filter==24.1
django-redis==5.4.0
django-reversion==5.0.12
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-rest-swagger==2.2.0
dnspython==2.6.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
kombu==5.3.5
msgpack==1.0.7
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.2
pyOpenSSL==24.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-stdnum==1.19
pytz==2024.1
pyyaml==6.0.1
redis==5.0.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
zxcvbn==4.4.28
isort==5.13.2
pylint==3.1.0
faker==24.1.0
